requirements.txt
Flask==2.3.2
line-bot-sdk==3.0.1
transformers==4.33.0
pydub==0.25.1
librosa==0.10.0.post2
soundfile==0.12.1
torch==2.0.1
python-dotenv==1.0.0
